Feat : 댓글 수정, 삭제 및 세션 로직 추가

+++ src/main/java/com/wonbin/practice/controller/CommentController.java
@@ -2,7 +2,9 @@ package com.wonbin.practice.controller;
 
 import com.wonbin.practice.dto.BoardDto;
 import com.wonbin.practice.dto.CommentDto;
+import com.wonbin.practice.dto.MemberDto;
 import com.wonbin.practice.service.CommentService;
+import com.wonbin.practice.service.MemberService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpSession;
@@ -13,6 +15,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
 @Tag(name = "comment")
 @Controller
 @RequiredArgsConstructor
@@ -20,13 +23,30 @@ import java.util.List;
 public class CommentController {
 
     private final CommentService commentService;
+    private final MemberService memberService;
 
-    @GetMapping("/{id}")
-    public ResponseEntity findByBoardId(@PathVariable Long id){
-        List<CommentDto> commentDtoList = commentService.findAll(id);
-        if(commentDtoList != null){
+    @GetMapping("/read/{commentId}")
+    public ResponseEntity findById(@PathVariable Long commentId, HttpSession session) {
+        String email = (String) session.getAttribute("loginEmail");
+        MemberDto memberDto = memberService.findByMemberEmail(email);
+        if (memberDto != null) {
+            CommentDto commentDto = commentService.findById(commentId);
+            if (commentDto != null) {
+                return new ResponseEntity<>(commentDto, HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("해당 댓글 조회 실패", HttpStatus.BAD_REQUEST);
+            }
+        } else {
+            return new ResponseEntity<>("로그인이 필요합니다.", HttpStatus.BAD_REQUEST);
+        }
+    }
+
+    @GetMapping("/{boardId}")
+    public ResponseEntity findByBoardId(@PathVariable Long boardId) {
+        List<CommentDto> commentDtoList = commentService.findAll(boardId);
+        if (commentDtoList != null) {
             return new ResponseEntity<>(commentDtoList, HttpStatus.OK);
-        }else {
+        } else {
             return new ResponseEntity<>(1, HttpStatus.BAD_REQUEST);
         }
     }
@@ -34,43 +54,52 @@ public class CommentController {
 
     @Operation(summary = "댓글 저장", description = "댓글을 저장합니다.")
     @PostMapping("/save/{boardId}")
-    public ResponseEntity save(@PathVariable Long boardId, @ModelAttribute CommentDto commentDto, HttpSession session){
+    public ResponseEntity save(@PathVariable Long boardId, @ModelAttribute CommentDto commentDto, HttpSession session) {
         System.out.println("commentDto = " + commentDto);
         commentDto.setBoardId(boardId);
-        String email =(String) session.getAttribute("loginEmail");
+        String email = (String) session.getAttribute("loginEmail");
         Long saveResult = commentService.save(commentDto, email);
-        if(saveResult != null){
+        if (saveResult > 0L) {
             // 작성 성공 후 댓글 목록 리턴
-//            List<CommentDto> commentDtoList = commentService.findAll(commentDto.getBoardId());
-//            return new ResponseEntity<>(commentDtoList, HttpStatus.OK);
             return new ResponseEntity<>(1, HttpStatus.OK);
-        }else {
-            return new ResponseEntity<>("댓글 저장 실패", HttpStatus.NOT_FOUND);
+        } else if (saveResult == -1L) {
+            return new ResponseEntity<>("로그인이 필요합니다.", HttpStatus.UNAUTHORIZED);
+        } else {
+            return new ResponseEntity<>("댓글 저장 실패", HttpStatus.BAD_REQUEST);
         }
     }
 
     @Operation(summary = "댓글 삭제", description = "댓글을 삭제합니다.")
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable Long id){
-
-        Long boardId = commentService.delete(id);
-        if(boardId == 0L){
-            System.out.println("널값");
+    public ResponseEntity delete(@PathVariable Long id, HttpSession session) {
+        String email = (String) session.getAttribute("loginEmail");
+        MemberDto memberDto = memberService.findByMemberEmail(email);
+        if (memberDto != null) {
+            Long boardId = commentService.delete(id);
+            if (boardId == 0L) {
+                return new ResponseEntity<>("댓글 삭제 실패", HttpStatus.BAD_REQUEST);
+            }
+            return new ResponseEntity<>("댓글 삭제 성공", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("로그인이 필요합니다.", HttpStatus.UNAUTHORIZED);
         }
-        return "redirect:/board/"+boardId;
     }
 
     @Operation(summary = "댓글 수정", description = "댓글을 수정합니다.")
-    @PostMapping("/update")
-    public ResponseEntity update(@ModelAttribute CommentDto commentDto){
+    @PostMapping("/update/{boardId}")
+    public ResponseEntity update(@ModelAttribute CommentDto commentDto, @PathVariable Long boardId, HttpSession session) {
         System.out.println(commentDto.toString());
-        Long saveResult = commentService.update(commentDto);
-        if(saveResult != null){
-            // 수정 성공 후 댓글 목록 리턴
-            List<CommentDto> commentDtoList = commentService.findAll(commentDto.getBoardId());
-            return new ResponseEntity<>(commentDtoList, HttpStatus.OK);
-        }else {
-            return new ResponseEntity<>("해당 댓글이 존재하지 않습니다.", HttpStatus.NOT_FOUND);
+        String email = (String) session.getAttribute("loginEmail");
+        MemberDto memberDto = memberService.findByMemberEmail(email);
+        if (memberDto != null) {
+            Long saveResult = commentService.update(commentDto, boardId, memberDto.getMemberEmail());
+            if (saveResult != null) {
+                return new ResponseEntity<>("댓글 수정에 성공했습니다.", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("해당 댓글이 존재하지 않습니다.", HttpStatus.NOT_FOUND);
+            }
+        } else {
+            return new ResponseEntity<>("로그인이 필요합니다.", HttpStatus.UNAUTHORIZED);
         }
     }
 }
